Fix script
The script, generate-theme.ts won't run.  There were a couple issues
identified.  This script is being used to generate a default theme,
it isn't actually part of the extension.  It's used to make sure the
extension gets published with the same theme every time.

Since it doesn't actually run in the extension host, we're getting
an error about loading the vscode module.  The solution is to NOT
use the generateTheme.default() method in favor of creating a new
Theme object inside the script.

There was a second problem in that the compile script wasn't allowing
the JSON module to be imported.  I found that the compile script
wasn't picking up the scripts/tsconfig.json file.  Fixed that,
reference https://www.npmjs.com/package/ts-node#configuration-options.

Lastly, since we need to import JSON to compile the script and to
compile the src/ folder, I added the resolveJsonModule setting to
the tsconfig.base.json file so that propagates down to the
tsconfig.json files in scripts/ and in src/.