Add One Dark Pro flattened theme
This commit adds a flattened version of the iconic and elegant One Dark
Pro theme, with an added emphasis on minimalism. It is inspired from the
Community Material Theme while maintaining the color scheme of One Dark
Pro. Tweaks to non-editor colors ensure that the user's focus is drawn
to the code instead of to other UI elements.

Additionally, this commit Closes #452 by correcting the typo in the enum
of the `package.json` file and adds the new "oneDarkProFlat" as an
option.