Prevent CD run on new branch
The create event is triggered by both new tags and new branches.
The workflow would be triggered and would fail when any new branch
was created on GitHub.  We band-aid this by putting a condition on
the deploy job, so it only runs if the thing being created is a tag.